backend/app/utils/sports_utils.py
"""
This module provides utility functions for fetching sports data.

References:
- https://github.com/swar/nba_api
- https://www.api-football.com/
- https://gist.github.com/nntrn/ee26cb2a0716de0947a0a4e9a157bc1c#list-of-nfl-api-endpoints
- https://rapidapi.com/Creativesdev/api/nfl-api-data
"""
import requests
import os

# --------------------------------------------------------------
# GLOBAL VARIABLES
# --------------------------------------------------------------



# --------------------------------------------------------------
# FUNCTIONS
# --------------------------------------------------------------

def fetch_sports_data() -> str:
    """
    Fetch sports data from the Sports API for chatbot responses

    :return: The sports data
    """
    try:
        return "Sports data"
    except Exception as e:
        raise RuntimeError(f"Error processing sports data: {e}")
