backend/requirements.txt
Flask
Flask-Cors
google-generativeai
SQLAlchemy
mysqlclient
pymysql
python-dotenv